+++ eotsmanager/cmd/eotsd/daemon/pop_test.go
@@ -1,6 +1,9 @@
 package daemon_test
 
 import (
+	"bytes"
+	"context"
+	"encoding/json"
 	"testing"
 
 	"github.com/babylonlabs-io/finality-provider/eotsmanager/cmd/eotsd/daemon"
@@ -59,8 +62,28 @@ func TestPoPValidBabySignEotsPk(t *testing.T) {
 func TestPoPVerify(t *testing.T) {
 	t.Parallel()
 	for _, pop := range popsToVerify {
-		valid, err := daemon.VerifyPopExport(pop)
+		valid, err := daemon.ValidPopExport(pop)
 		require.NoError(t, err)
 		require.True(t, valid)
+	}
+}
+
+func TestPoPValidate(t *testing.T) {
+	t.Parallel()
+	validateCmd := daemon.NewPopValidateExportCmd()
+
+	for _, pop := range popsToVerify {
+		jsonString, err := json.MarshalIndent(pop, "", "  ")
+		require.NoError(t, err)
+
+		writer := bytes.NewBuffer([]byte{})
+		validateCmd.SetOutput(writer)
+
+		validateCmd.SetArgs([]string{string(jsonString)})
+
+		err = validateCmd.ExecuteContext(context.Background())
+		require.NoError(t, err)
+
+		require.Equal(t, writer.String(), "Proof of Possession is valid!\n")
 	}
 }